ai/OCR_V2_inference/requirements.txt
# VietOCR API Service Requirements

# FastAPI and related
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# VietOCR
vietocr==0.3.12

# PyTorch (required by VietOCR)
torch>=1.9.0
torchvision>=0.10.0

# Image Processing (Pillow 9.x for VietOCR compatibility)
Pillow==9.5.0

# Utilities
numpy>=1.21.0
gdown==5.1.0

# CORS
python-dotenv==1.0.0
